[flutter_tools] split build directories from build info
