+++ packages/flutter_tools/lib/src/android/application_package.dart
@@ -10,132 +10,18 @@ import 'package:meta/meta.dart';
 import 'package:process/process.dart';
 import 'package:xml/xml.dart';
 
-import 'android/android_sdk.dart';
-import 'android/gradle.dart';
-import 'base/common.dart';
-import 'base/context.dart';
-import 'base/file_system.dart';
-import 'base/io.dart';
-import 'base/logger.dart';
-import 'base/process.dart';
-import 'base/user_messages.dart';
-import 'build_directories.dart';
-import 'build_info.dart';
-import 'fuchsia/application_package.dart';
-import 'globals.dart' as globals;
-import 'ios/plist_parser.dart';
-import 'linux/application_package.dart';
-import 'macos/application_package.dart';
-import 'project.dart';
-import 'tester/flutter_tester.dart';
-import 'web/web_device.dart';
-import 'windows/application_package.dart';
-
-class ApplicationPackageFactory {
-  ApplicationPackageFactory({
-    @required AndroidSdk androidSdk,
-    @required ProcessManager processManager,
-    @required Logger logger,
-    @required UserMessages userMessages,
-    @required FileSystem fileSystem,
-  }) : _androidSdk = androidSdk,
-       _processManager = processManager,
-       _logger = logger,
-       _userMessages = userMessages,
-       _fileSystem = fileSystem,
-       _processUtils = ProcessUtils(logger: logger, processManager: processManager);
-
-  static ApplicationPackageFactory get instance => context.get<ApplicationPackageFactory>();
-
-  final AndroidSdk _androidSdk;
-  final ProcessManager _processManager;
-  final Logger _logger;
-  final ProcessUtils _processUtils;
-  final UserMessages _userMessages;
-  final FileSystem _fileSystem;
-
-  Future<ApplicationPackage> getPackageForPlatform(
-    TargetPlatform platform, {
-    BuildInfo buildInfo,
-    File applicationBinary,
-  }) async {
-    switch (platform) {
-      case TargetPlatform.android:
-      case TargetPlatform.android_arm:
-      case TargetPlatform.android_arm64:
-      case TargetPlatform.android_x64:
-      case TargetPlatform.android_x86:
-        if (applicationBinary == null) {
-          return AndroidApk.fromAndroidProject(
-            FlutterProject.current().android,
-            processManager: _processManager,
-            processUtils: _processUtils,
-            logger: _logger,
-            androidSdk: _androidSdk,
-            userMessages: _userMessages,
-            fileSystem: _fileSystem,
-          );
-        }
-        return AndroidApk.fromApk(
-          applicationBinary,
-          processManager: _processManager,
-          logger: _logger,
-          androidSdk: _androidSdk,
-          userMessages: _userMessages,
-        );
-      case TargetPlatform.ios:
-        return applicationBinary == null
-            ? await IOSApp.fromIosProject(FlutterProject.current().ios, buildInfo)
-            : IOSApp.fromPrebuiltApp(applicationBinary);
-      case TargetPlatform.tester:
-        return FlutterTesterApp.fromCurrentDirectory(globals.fs);
-      case TargetPlatform.darwin_x64:
-        return applicationBinary == null
-            ? MacOSApp.fromMacOSProject(FlutterProject.current().macos)
-            : MacOSApp.fromPrebuiltApp(applicationBinary);
-      case TargetPlatform.web_javascript:
-        if (!FlutterProject.current().web.existsSync()) {
-          return null;
-        }
-        return WebApplicationPackage(FlutterProject.current());
-      case TargetPlatform.linux_x64:
-      case TargetPlatform.linux_arm64:
-        return applicationBinary == null
-            ? LinuxApp.fromLinuxProject(FlutterProject.current().linux)
-            : LinuxApp.fromPrebuiltApp(applicationBinary);
-      case TargetPlatform.windows_x64:
-        return applicationBinary == null
-            ? WindowsApp.fromWindowsProject(FlutterProject.current().windows)
-            : WindowsApp.fromPrebuiltApp(applicationBinary);
-      case TargetPlatform.fuchsia_arm64:
-      case TargetPlatform.fuchsia_x64:
-        return applicationBinary == null
-            ? FuchsiaApp.fromFuchsiaProject(FlutterProject.current().fuchsia)
-            : FuchsiaApp.fromPrebuiltApp(applicationBinary);
-      case TargetPlatform.windows_uwp_x64:
-        throw UnsupportedError('Cannot build for windows_uwp_x64');
-    }
-    assert(platform != null);
-    return null;
-  }
-}
-
-abstract class ApplicationPackage {
-  ApplicationPackage({ @required this.id })
-    : assert(id != null);
-
-  /// Package ID from the Android Manifest or equivalent.
-  final String id;
-
-  String get name;
-
-  String get displayName => name;
-
-  File get packagesFile => null;
-
-  @override
-  String toString() => displayName ?? id;
-}
+import '../application_package.dart';
+import '../base/common.dart';
+import '../base/file_system.dart';
+import '../base/io.dart';
+import '../base/logger.dart';
+import '../base/process.dart';
+import '../base/user_messages.dart';
+import '../build_directories.dart';
+import '../globals.dart' as globals;
+import '../project.dart';
+import 'android_sdk.dart';
+import 'gradle.dart';
 
 /// An application package created from an already built Android APK.
 class AndroidApk extends ApplicationPackage {
@@ -323,170 +209,6 @@ class AndroidApk extends ApplicationPackage {
   String get name => file.basename;
 }
 
-/// Tests whether a [Directory] is an iOS bundle directory.
-bool _isBundleDirectory(Directory dir) => dir.path.endsWith('.app');
-
-abstract class IOSApp extends ApplicationPackage {
-  IOSApp({@required String projectBundleId}) : super(id: projectBundleId);
-
-  /// Creates a new IOSApp from an existing app bundle or IPA.
-  factory IOSApp.fromPrebuiltApp(FileSystemEntity applicationBinary) {
-    final FileSystemEntityType entityType = globals.fs.typeSync(applicationBinary.path);
-    if (entityType == FileSystemEntityType.notFound) {
-      globals.printError(
-          'File "${applicationBinary.path}" does not exist. Use an app bundle or an ipa.');
-      return null;
-    }
-    Directory bundleDir;
-    if (entityType == FileSystemEntityType.directory) {
-      final Directory directory = globals.fs.directory(applicationBinary);
-      if (!_isBundleDirectory(directory)) {
-        globals.printError('Folder "${applicationBinary.path}" is not an app bundle.');
-        return null;
-      }
-      bundleDir = globals.fs.directory(applicationBinary);
-    } else {
-      // Try to unpack as an ipa.
-      final Directory tempDir = globals.fs.systemTempDirectory.createTempSync('flutter_app.');
-      globals.os.unzip(globals.fs.file(applicationBinary), tempDir);
-      final Directory payloadDir = globals.fs.directory(
-        globals.fs.path.join(tempDir.path, 'Payload'),
-      );
-      if (!payloadDir.existsSync()) {
-        globals.printError(
-            'Invalid prebuilt iOS ipa. Does not contain a "Payload" directory.');
-        return null;
-      }
-      try {
-        bundleDir = payloadDir.listSync().whereType<Directory>().singleWhere(_isBundleDirectory);
-      } on StateError {
-        globals.printError(
-            'Invalid prebuilt iOS ipa. Does not contain a single app bundle.');
-        return null;
-      }
-    }
-    final String plistPath = globals.fs.path.join(bundleDir.path, 'Info.plist');
-    if (!globals.fs.file(plistPath).existsSync()) {
-      globals.printError('Invalid prebuilt iOS app. Does not contain Info.plist.');
-      return null;
-    }
-    final String id = globals.plistParser.getValueFromFile(
-      plistPath,
-      PlistParser.kCFBundleIdentifierKey,
-    );
-    if (id == null) {
-      globals.printError('Invalid prebuilt iOS app. Info.plist does not contain bundle identifier');
-      return null;
-    }
-
-    return PrebuiltIOSApp(
-      bundleDir: bundleDir,
-      bundleName: globals.fs.path.basename(bundleDir.path),
-      projectBundleId: id,
-    );
-  }
-
-  static Future<IOSApp> fromIosProject(IosProject project, BuildInfo buildInfo) {
-    if (!globals.platform.isMacOS) {
-      return null;
-    }
-    if (!project.exists) {
-      // If the project doesn't exist at all the current hint to run flutter
-      // create is accurate.
-      return null;
-    }
-    if (!project.xcodeProject.existsSync()) {
-      globals.printError('Expected ios/Runner.xcodeproj but this file is missing.');
-      return null;
-    }
-    if (!project.xcodeProjectInfoFile.existsSync()) {
-      globals.printError('Expected ios/Runner.xcodeproj/project.pbxproj but this file is missing.');
-      return null;
-    }
-    return BuildableIOSApp.fromProject(project, buildInfo);
-  }
-
-  @override
-  String get displayName => id;
-
-  String get simulatorBundlePath;
-
-  String get deviceBundlePath;
-
-  /// Directory used by ios-deploy to store incremental installation metadata for
-  /// faster second installs.
-  Directory get appDeltaDirectory;
-}
-
-class BuildableIOSApp extends IOSApp {
-  BuildableIOSApp(this.project, String projectBundleId, String hostAppBundleName)
-    : _hostAppBundleName = hostAppBundleName,
-      super(projectBundleId: projectBundleId);
-
-  static Future<BuildableIOSApp> fromProject(IosProject project, BuildInfo buildInfo) async {
-    final String projectBundleId = await project.productBundleIdentifier(buildInfo);
-    final String hostAppBundleName = await project.hostAppBundleName(buildInfo);
-    return BuildableIOSApp(project, projectBundleId, hostAppBundleName);
-  }
-
-  final IosProject project;
-
-  final String _hostAppBundleName;
-
-  @override
-  String get name => _hostAppBundleName;
-
-  @override
-  String get simulatorBundlePath => _buildAppPath('iphonesimulator');
-
-  @override
-  String get deviceBundlePath => _buildAppPath('iphoneos');
-
-  @override
-  Directory get appDeltaDirectory => globals.fs.directory(globals.fs.path.join(getIosBuildDirectory(), 'app-delta'));
-
-  // Xcode uses this path for the final archive bundle location,
-  // not a top-level output directory.
-  // Specifying `build/ios/archive/Runner` will result in `build/ios/archive/Runner.xcarchive`.
-  String get archiveBundlePath
-    => globals.fs.path.join(getIosBuildDirectory(), 'archive', globals.fs.path.withoutExtension(_hostAppBundleName));
-
-  // The output xcarchive bundle path `build/ios/archive/Runner.xcarchive`.
-  String get archiveBundleOutputPath =>
-      globals.fs.path.setExtension(archiveBundlePath, '.xcarchive');
-
-  String get ipaOutputPath =>
-      globals.fs.path.join(getIosBuildDirectory(), 'ipa');
-
-  String _buildAppPath(String type) {
-    return globals.fs.path.join(getIosBuildDirectory(), type, _hostAppBundleName);
-  }
-}
-
-class PrebuiltIOSApp extends IOSApp {
-  PrebuiltIOSApp({
-    this.bundleDir,
-    this.bundleName,
-    @required String projectBundleId,
-  }) : super(projectBundleId: projectBundleId);
-
-  final Directory bundleDir;
-  final String bundleName;
-
-  @override
-  final Directory appDeltaDirectory = null;
-
-  @override
-  String get name => bundleName;
-
-  @override
-  String get simulatorBundlePath => _bundlePath;
-
-  @override
-  String get deviceBundlePath => _bundlePath;
-
-  String get _bundlePath => bundleDir.path;
-}
 
 class _Entry {
   _Element parent;
