+++ packages/flutter_tools/lib/src/bundle.dart
@@ -14,6 +14,7 @@ import 'base/common.dart';
 import 'base/config.dart';
 import 'base/file_system.dart';
 import 'base/logger.dart';
+import 'build_directories.dart';
 import 'build_info.dart';
 import 'build_system/build_system.dart';
 import 'build_system/depfile.dart';
