+++ packages/flutter_tools/lib/src/devfs.dart
@@ -17,7 +17,7 @@ import 'base/io.dart';
 import 'base/logger.dart';
 import 'base/net.dart';
 import 'base/os.dart';
-import 'build_info.dart';
+import 'build_directories.dart';
 import 'compile.dart';
 import 'convert.dart' show base64, utf8;
 import 'vmservice.dart';
