+++ packages/flutter_tools/lib/src/ios/mac.dart
@@ -16,6 +16,7 @@ import '../base/logger.dart';
 import '../base/process.dart';
 import '../base/project_migrator.dart';
 import '../base/utils.dart';
+import '../build_directories.dart';
 import '../build_info.dart';
 import '../cache.dart';
 import '../flutter_manifest.dart';
