+++ packages/flutter_tools/lib/src/isolated/devfs_web.dart
@@ -26,6 +26,7 @@ import '../base/io.dart';
 import '../base/logger.dart';
 import '../base/net.dart';
 import '../base/platform.dart';
+import '../build_directories.dart';
 import '../build_info.dart';
 import '../bundle.dart';
 import '../cache.dart';
