+++ packages/flutter_tools/lib/src/linux/application_package.dart
@@ -8,6 +8,7 @@ import 'package:meta/meta.dart';
 
 import '../application_package.dart';
 import '../base/file_system.dart';
+import '../build_directories.dart';
 import '../build_info.dart';
 import '../cmake.dart';
 import '../globals.dart' as globals;
