+++ packages/flutter_tools/lib/src/run_cold.dart
@@ -10,7 +10,7 @@ import 'package:meta/meta.dart';
 
 import 'base/common.dart';
 import 'base/file_system.dart';
-import 'build_info.dart';
+import 'build_directories.dart';
 import 'device.dart';
 import 'globals.dart' as globals;
 import 'resident_devtools_handler.dart';
