+++ packages/flutter_tools/lib/src/run_hot.dart
@@ -17,6 +17,7 @@ import 'base/file_system.dart';
 import 'base/logger.dart';
 import 'base/platform.dart';
 import 'base/utils.dart';
+import 'build_directories.dart';
 import 'build_info.dart';
 import 'bundle.dart';
 import 'compile.dart';
