+++ packages/flutter_tools/lib/src/test/test_compiler.dart
@@ -10,6 +10,7 @@ import 'package:meta/meta.dart';
 
 import '../artifacts.dart';
 import '../base/file_system.dart';
+import '../build_directories.dart';
 import '../build_info.dart';
 import '../bundle.dart';
 import '../compile.dart';
