+++ packages/flutter_tools/lib/src/windows/application_package.dart
@@ -9,6 +9,7 @@ import 'package:meta/meta.dart';
 import '../application_package.dart';
 import '../base/file_system.dart';
 import '../base/utils.dart';
+import '../build_directories.dart';
 import '../build_info.dart';
 import '../cmake.dart';
 import '../globals.dart' as globals;
